+++ docs/CHANGELOG.md
@@ -71,6 +71,7 @@ about the REST API.
 * User Experience:
   * When reading text, you can click on "Show All" and "Translations" to click the checkbox.
   * When changing a word status, the displayed messaged is now more informative: before database change, success message, now: waiting screen, tentative database change, result message (error or success). Under the hood, we are transitionning from pure PHP to AJAX.
+  * It is faster to exit a text page: audio and reading position saves are now asynchronous.
 * Annotations "order" are now identical to ``Ti2Order`` (in ``textitems2``). It makes management easier, but users may need to reload the annotations once.
 * Updated composer dependencies (up-to-date 2023-11-23).